arm: imx8qm: check magic_num
The feature to let M41 communicate with XEN is not ready,
so we have to add code to tell xen whether it is in
U-Boot stage or Linux kernel stage.

If in U-Boot stage, not shutdown the power domains, otherwise
M41 will lose display because it not aware of power domain shutdown
now.

In Linux kernel stage, M41 is aware of shutdown, because there is rpmsg
communication between Android Auto and M41.

So ingore_power_off is only valid in U-Boot stage, the magic number
is written by U-Boot and cleared before booting Linux Kernel.

Signed-off-by: Peng Fan <peng.fan@nxp.com>
Reviewed-by: Flynn xu <flynn.xu@nxp.com>