imx8qm: implement system reset
We created scfw partitions for DomU android, before reboot,
need to free them, the best practice is that scfw could
free all the child's child partition, that will be done
in scfw side. But currently we need this patch to let m4
work well when xen ctrl + A force reboot when android auto
is running.

Signed-off-by: Peng Fan <peng.fan@nxp.com>